Rename FilterableSideNav to SideNav
