+++ src/implementations/react/src/adapters/GlobalNav/SideNav/SideNavAdapter.test.js
@@ -1,4 +1,3 @@
-
 import { mount } from 'enzyme';
 import * as HIG from 'hig-vanilla';
 import React from 'react';
