--- src/implementations/react/src/elements/components/GlobalNav/FilterableSideNav/index.js
@@ -1,3 +1,0 @@
-import FilterableSideNav from './FilterableSideNav';
-
-export default FilterableSideNav;
