+++ src/implementations/react/src/elements/components/GlobalNav/SideNav/Module.js
