+++ src/implementations/react/src/elements/components/GlobalNav/SideNav/ModuleCollapse.js
