+++ src/implementations/react/src/elements/components/GlobalNav/SideNav/index.js
@@ -1,0 +1,3 @@
+import SideNav from './SideNav';
+
+export default SideNav;
