+++ src/implementations/react/src/elements/components/GlobalNav/SideNav/model.js
